+++ src/test/java/com/hazelcast/wm/test/jetty/ConcurrentRequestTest.java
@@ -13,10 +13,13 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.hazelcast.wm.test;
+package com.hazelcast.wm.test.jetty;
 
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.wm.test.AbstractWebFilterTest;
+import com.hazelcast.wm.test.ServletContainer;
+import com.hazelcast.wm.test.jetty.JettyServer;
 import org.apache.http.client.CookieStore;
 import org.apache.http.impl.client.BasicCookieStore;
 import org.junit.Test;
