+++ src/test/java/com/hazelcast/wm/test/jetty/DeferredWriteClusterTest.java
@@ -13,18 +13,26 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.hazelcast.wm.test;
+package com.hazelcast.wm.test.jetty;
 
+import com.hazelcast.config.FileSystemXmlConfig;
 import com.hazelcast.core.Hazelcast;
-import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.map.IMap;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.web.SessionState;
+import com.hazelcast.wm.test.AbstractWebFilterTest;
+import com.hazelcast.wm.test.ServletContainer;
+import com.hazelcast.wm.test.jetty.JettyServer;
 import org.apache.http.client.CookieStore;
 import org.apache.http.impl.client.BasicCookieStore;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
+
+import java.io.File;
 
 import static org.junit.Assert.assertEquals;
 
@@ -37,10 +45,10 @@ import static org.junit.Assert.assertEquals;
  */
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
-public class DisconnectedHazelcastTest extends AbstractWebFilterTest {
+public class DeferredWriteClusterTest extends AbstractWebFilterTest {
 
-    public DisconnectedHazelcastTest() {
-        super("node1-client.xml", "node2-client.xml");
+    public DeferredWriteClusterTest() {
+        super("node1-client-deferred.xml", "node2-client-deferred.xml");
     }
 
     @Before
@@ -76,10 +84,17 @@ public class DisconnectedHazelcastTest extends AbstractWebFilterTest {
             // So, we should copy current information from test to current context.
             cc.copyFrom(this);
         }
+        // Clear map
+//        IMap<String, Object> map = hz.getMap(DEFAULT_MAP_NAME);
+//        map.clear();
     }
 
     @Override
     protected void ensureInstanceIsUp() throws Exception {
+        if (isInstanceNotActive(hz)) {
+            hz = Hazelcast.newHazelcastInstance(
+                    new FileSystemXmlConfig(new File(sourceDir + "/WEB-INF/", "hazelcast.xml")));
+        }
         if (serverXml1 != null) {
             if (server1 == null) {
                 serverPort1 = availablePort();
@@ -98,32 +113,41 @@ public class DisconnectedHazelcastTest extends AbstractWebFilterTest {
         }
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = 20000)
     public void test_setAttribute() throws Exception {
         CookieStore cookieStore = new BasicCookieStore();
-        executeRequest("write", serverPort1, cookieStore);
-        assertEquals("value", executeRequest("read", serverPort1, cookieStore));
-        assertEquals("null", executeRequest("read", serverPort2, cookieStore));
-    }
-
-    @Test(timeout = 60000)
-    public void test_getAttribute() throws Exception {
-        CookieStore cookieStore = new BasicCookieStore();
-        assertEquals("null", executeRequest("read", serverPort1, cookieStore));
-        executeRequest("write", serverPort1, cookieStore);
-        assertEquals("value", executeRequest("read", serverPort1, cookieStore));
-        assertEquals("value", executeRequest("readIfExist", serverPort1, cookieStore));
-        assertEquals("null", executeRequest("read", serverPort2, cookieStore));
-        HazelcastInstance hz = Hazelcast.newHazelcastInstance();
-        assertClusterSizeEventually(1, hz);
-        Thread.sleep(9000);
-        executeRequest("write", serverPort1, cookieStore);
+        executeRequest("setAttribute?key=value&key2=value2&key2=value22", serverPort1, cookieStore);
+        IMap<Object, Object> map = hz.getMap(DEFAULT_MAP_NAME);
+        assertEquals(1, map.size());
+        SessionState sessionState = (SessionState) map.get(getHazelcastSessionId(cookieStore));
+        assertEquals(2, sessionState.getAttributes().size());
         assertEquals("value", executeRequest("read", serverPort1, cookieStore));
         assertEquals("value", executeRequest("read", serverPort2, cookieStore));
+        assertEquals(1, map.size());
+        assertEquals("value", executeRequest("read", serverPort2, cookieStore));
+        assertEquals("true", executeRequest("update", serverPort2, cookieStore));
+        assertEquals("value-updated", executeRequest("read", serverPort2, cookieStore));
+        assertEquals("value-updated", executeRequest("read", serverPort1, cookieStore));
+    }
+    @Test
+    public void testMultipleGetSession() throws Exception {
+        CookieStore cookieStore = new BasicCookieStore();
+        assertEquals("value", executeRequest("multiplesession", serverPort1, cookieStore));
+    }
+
+    @Test(timeout = 20000)
+    public void test_setThenGetAttribute() throws Exception {
+        CookieStore cookieStore = new BasicCookieStore();
+        assertEquals("value", executeRequest("setGet", serverPort1, cookieStore));
     }
 
     @Override
     protected ServletContainer getServletContainer(int port, String sourceDir, String serverXml) throws Exception {
         return new JettyServer(port, sourceDir, serverXml);
+    }
+
+    @After
+    public void shutdown() throws Exception {
+        teardownClass();
     }
 }
