+++ src/test/java/com/hazelcast/wm/test/jetty/WebFilterSessionCleanupTest.java
@@ -13,11 +13,16 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.hazelcast.wm.test;
+package com.hazelcast.wm.test.jetty;
 
 import com.hazelcast.map.IMap;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.wm.test.AbstractWebFilterTest;
+import com.hazelcast.wm.test.DelegatedRunWith;
+import com.hazelcast.wm.test.ServletContainer;
+import com.hazelcast.wm.test.WebTestRunner;
+import com.hazelcast.wm.test.jetty.JettyServer;
 import org.apache.http.client.CookieStore;
 import org.apache.http.impl.client.BasicCookieStore;
 import org.junit.Test;
