+++ src/test/java/com/hazelcast/wm/test/tomcat/TomcatKeepRemoteActiveTest.java
@@ -1,6 +1,10 @@
-package com.hazelcast.wm.test;
+package com.hazelcast.wm.test.tomcat;
 
 import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.wm.test.DelegatedRunWith;
+import com.hazelcast.wm.test.KeepRemoteActiveTest;
+import com.hazelcast.wm.test.ServletContainer;
+import com.hazelcast.wm.test.WebTestRunner;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -11,7 +15,7 @@ import java.util.Collection;
 @RunWith(WebTestRunner.class)
 @DelegatedRunWith(Parameterized.class)
 @Category(QuickTest.class)
-public class JettyKeepRemoteActiveTest extends KeepRemoteActiveTest {
+public class TomcatKeepRemoteActiveTest extends KeepRemoteActiveTest {
 
     @Parameterized.Parameters(name = "Executing: {0}")
     public static Collection<Object[]> parameters() {
@@ -21,13 +25,13 @@ public class JettyKeepRemoteActiveTest extends KeepRemoteActiveTest {
         );
     }
 
-    public JettyKeepRemoteActiveTest(String serverXml1, String serverXml2) {
+    public TomcatKeepRemoteActiveTest(String serverXml1, String serverXml2) {
         super(serverXml1,serverXml2);
     }
 
     @Override
     protected ServletContainer getServletContainer(int port, String sourceDir, String serverXml) throws Exception {
-        return new JettyServer(port,sourceDir,serverXml);
+        return new TomcatServer(port,sourceDir,serverXml);
     }
 
 }
